+++ checkout-ui-custom/src/checkout6-custom.js
@@ -1,9 +1,11 @@
 import CartController from './controller/CartController';
+import AddressController from './controller/AddressController';
 import ViewController from './controller/ViewController';
 import FormController from './controller/FormController';
 import CollectController from './controller/CollectController';
 
 CartController.init();
+AddressController.init();
 ViewController.init();
 FormController.init();
 CollectController.init();