+++ checkout-ui-custom/src/controller/CartController.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-undef */
 import { STEPS, TIMEOUT_500 } from '../utils/const';
 
 const CartController = (() => {
@@ -39,19 +38,17 @@ const CartController = (() => {
   };
 
   const runCustomization = () => {
-    if (window.location.hash === STEPS.SHIPPING || window.location.hash === STEPS.PAYMENT) {
-      if (typeof (setAppConfiguration) !== 'undefined') {
-        // eslint-disable-next-line no-undef
-        setAppConfiguration(config);
-      }
-
-      setTimeout(() => {
-        if (vtexjs.checkout.orderForm) {
-          const { items } = vtexjs.checkout.orderForm;
-          getCategories(items);
-        }
-      }, TIMEOUT_500);
+    if (typeof (setAppConfiguration) !== 'undefined') {
+      // eslint-disable-next-line no-undef
+      setAppConfiguration(config);
     }
+
+    setTimeout(() => {
+      if (window.vtexjs.checkout.orderForm) {
+        const { items } = window.vtexjs.checkout.orderForm;
+        getCategories(items);
+      }
+    }, TIMEOUT_500);
   };
 
   // EVENTS SUBSCRIPTION