+++ checkout-ui-custom/src/templates/index.js
@@ -5,6 +5,7 @@ import RICAForm from './RICAForm';
 import InputError from './InputError';
 import MixedProducts from './MixedProducts';
 import AddressForm, { SuburbField, PickupComplementField } from './AddressForm';
+import AlertBox from './AlertBox';
 
 export {
   FurnitureForm,
@@ -14,5 +15,6 @@ export {
   MixedProducts,
   AddressForm,
   SuburbField,
+  AlertBox,
   PickupComplementField
 };