+++ checkout-ui-custom/src/utils/functions.js
@@ -10,7 +10,7 @@ const getShippingData = async (addressName, fields) => {
   };
 
   const response = await fetch(
-    '/custom-api/masterdata/addresses/' + `${fields}&_where=addressName=${addressName}&timestamp=${Date.now()}`,
+    `/custom-api/masterdata/addresses/${fields}&_where=addressName=${addressName}&timestamp=${Date.now()}`,
     options
   )
     .then((res) => res.json())
@@ -23,35 +23,37 @@ const getShippingData = async (addressName, fields) => {
   return data;
 };
 
-const saveAddress = async (fields) => {
+const saveAddress = async (fields = {}) => {
   let path;
-  const { email } = vtexjs.checkout.orderForm.clientProfileData;
-  const { address } = vtexjs.checkout.orderForm.shippingData;
+  const { email } = window.vtexjs.checkout.orderForm.clientProfileData;
+  const { address } = window.vtexjs.checkout.orderForm.shippingData;
+
+  if (!address) return;
 
   // AD already exists (?)
-  const savedAddress = await getShippingData(address.addressId, '?_fields=id');
+  const savedAddress = address?.addressId ? await getShippingData(address.addressId, '?_fields=id') : {};
 
-  if (savedAddress.id) {
+  if (savedAddress?.id) {
     path = `/custom-api/masterdata/address/${savedAddress.id}`;
   } else {
     path = '/custom-api/masterdata/addresses';
   }
 
-  // Importante respetar el order de address para no sobreescribir receiver, complement y neighborhood
-  const body = {
+  // Importante respetar el orden de address para no sobreescribir receiver, complement y neighborhood
+  const newAddress = {
     userId: email,
     ...address,
     ...fields
   };
 
   if (!savedAddress.id) {
-    body.addressName = address.addressId;
+    newAddress.addressName = address.addressId;
   }
 
   const options = {
     method: 'PATCH',
     headers: { 'Cache-Control': 'no-cache' },
-    body: JSON.stringify(body)
+    body: JSON.stringify(newAddress)
   };
 
   await fetch(path, options)
@@ -64,40 +66,47 @@ const saveAddress = async (fields) => {
     .catch((error) => console.log(error));
 };
 
-const setMasterdataFields = async (completeFurnitureForm, completeTVIDForm, tries = 0) => {
-  const { addressId } = vtexjs.checkout.orderForm.shippingData.address;
-  const fields = '?_fields=companyBuilding,furnitureReady,buildingType,'
-    + 'parkingDistance,deliveryFloor,liftOrStairs,hasSufficientSpace,assembleFurniture,tvID';
+const setMasterdataFields = async (completeFurnitureForm, completeTVIDForm, tries = 1) => {
+  /* Data will only be searched and set if any of the custom fields for TFG are displayed. */
+  if (completeFurnitureForm || completeTVIDForm) {
+    const { address } = window.vtexjs.checkout.orderForm.shippingData;
 
-  const customShippingInfo = await getShippingData(addressId, fields);
+    /* Setting Masterdata custom fields */
+    const fields = '?_fields=buildingType,parkingDistance,deliveryFloor,liftOrStairs,hasSufficientSpace'
+      + ',assembleFurniture,tvID';
 
-  if (customShippingInfo && !jQuery.isEmptyObject(customShippingInfo)) {
-    if (completeFurnitureForm) {
-      $('#tfg-building-type').val(customShippingInfo.buildingType);
-      $('#tfg-parking-distance').val(customShippingInfo.parkingDistance);
-      $('#tfg-delivery-floor').val(customShippingInfo.deliveryFloor);
-      if ($('#tfg-delivery-floor').val() === 'Ground') {
-        $('#tfg-lift-stairs').attr('disabled', 'disabled');
-      } else {
-        $('#tfg-lift-stairs').val(customShippingInfo.liftOrStairs);
+    const shippingData = await getShippingData(address.addressId, fields);
+
+    if (shippingData && !jQuery.isEmptyObject(shippingData)) {
+      /* Setting furniture form values */
+      if (completeFurnitureForm) {
+        $('#tfg-building-type').val(shippingData.buildingType);
+        $('#tfg-parking-distance').val(shippingData.parkingDistance);
+        $('#tfg-delivery-floor').val(shippingData.deliveryFloor);
+        if ($('#tfg-delivery-floor').val() === 'Ground') {
+          $('#tfg-lift-stairs').attr('disabled', 'disabled');
+        } else {
+          $('#tfg-lift-stairs').val(shippingData.liftOrStairs);
+        }
+        $('#tfg-sufficient-space').prop('checked', shippingData.hasSufficientSpace);
+        $('#tfg-assemble-furniture').prop('checked', shippingData.assembleFurniture);
       }
-      $('#tfg-sufficient-space').prop('checked', customShippingInfo.hasSufficientSpace);
-      $('#tfg-assemble-furniture').prop('checked', customShippingInfo.assembleFurniture);
-    }
 
-    if (completeTVIDForm) {
-      $('#tfg-tv-licence').val(customShippingInfo.tvID);
+      /* Setting TV form values */
+      if (completeTVIDForm) {
+        $('#tfg-tv-licence').val(shippingData.tvID);
+      }
+    } else if (tries <= 5) {
+      setTimeout(() => {
+        setMasterdataFields(completeFurnitureForm, completeTVIDForm, (tries += 1));
+      }, 3000);
     }
-  } else if (tries <= 5) {
-    setTimeout(() => {
-      setMasterdataFields(completeFurnitureForm, completeTVIDForm, (tries += 1));
-    }, 3000);
   }
 };
 
 // Functions to manage CustomData
 const checkoutGetCustomData = (appId) => {
-  const { customData } = vtexjs.checkout.orderForm;
+  const { customData } = window.vtexjs.checkout.orderForm;
   let fields = {};
 
   if (customData && customData.customApps && customData.customApps.length > 0) {
@@ -112,7 +121,7 @@ const checkoutGetCustomData = (appId) => {
 };
 
 const checkoutSendCustomData = (appId, customData) => {
-  const { orderFormId } = vtexjs.checkout.orderForm;
+  const { orderFormId } = window.vtexjs.checkout.orderForm;
 
   return $.ajax({
     type: 'PUT',
@@ -126,21 +135,21 @@ const checkoutSendCustomData = (appId, customData) => {
 const setRicaFields = (getDataFrom = 'customApps') => {
   let ricaFields;
 
-  if (getDataFrom === 'customApps') {
-    ricaFields = checkoutGetCustomData(RICA_APP);
-  } else if (getDataFrom === 'shippingAddress') {
-    const { address } = vtexjs.checkout.orderForm.shippingData;
+  if (getDataFrom === 'shippingAddress') {
+    const { address } = window.vtexjs.checkout.orderForm.shippingData;
 
     ricaFields = {
       idOrPassport: '',
       sameAddress: 'true',
-      fullName: address.receiverName,
+      fullName: address.receiverName || $('#ship-receiverName').val(),
       streetAddress: `${address.street}, ${address.number}`,
       suburb: address.neighborhood,
       city: address.city,
       postalCode: address.postalCode,
       province: address.state
     };
+  } else if (getDataFrom === 'customApps') {
+    ricaFields = checkoutGetCustomData(RICA_APP);
   }
 
   if (ricaFields && !jQuery.isEmptyObject(ricaFields)) {
@@ -159,10 +168,8 @@ const setRicaFields = (getDataFrom = 'customApps') => {
 
 // Random Functions
 const addBorderTop = (elementClass) => {
-  if ($(elementClass).length > 1) {
-    $(elementClass).addClass('custom-step-border');
-    $(elementClass).last().addClass('last-custom-step-border');
-  }
+  $(elementClass).addClass('custom-step-border');
+  $(elementClass).last().addClass('last-custom-step-border');
 };
 
 const waitAndResetLocalStorage = () => {