+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.1.4] - 2022-06-03
+### Fixed
+- fixed problem with field saving
+- design changed to more securely control receiver and complement field
+- fixed bugs with native behaviour
+
 ## [0.1.3] - 2022-05-31
 ### Added
 - added receiver, complement on collect step
@@ -14,6 +20,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.1.2] - 2022-XX-XX
 ### Added
+- Message when address is saved
+- Hide subheader on furniture case
+- Icons on custom messages
+
+### Changed
+- New CI: moved styles to "tfg-custom-checkout" repository
+- Validation fields
+
+### Fixed
+- Fixed bug saving custom fields in Masterdata (receiver, complement, neighborhood, company/building)
+- Not continue to payment when mandatory fields are empty and redirect to form
+- Error on lift or stairs label
+
+## [0.1.2] - 2022-XX-XX
+
+### Added
+
 - added receiver, complement and suburb with validation and save
 - added google autocomplete view
 - added complement with intl-tel-input
@@ -31,7 +54,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Functionality associated with RICA
 - Functionality associated with Furniture products
 - Functionality associated with Mixed products
+
 ### Fixed
 
 - Fixed issues after test all the cases
-